Comments
Added comments. Also added a circle to be drawn and some error handling

+++ include/ElectroSim/Electron.hpp
@@ -5,8 +5,8 @@
 
 class Electron : public Particle {
 public:
-Electron(int x, int y, int radius);
-void render();
+    Electron(float x, float y, float radius);
+    void render();
 };
 
 #endif // ifndef ELECTRON_HPP
