+++ src/Proton.cpp
@@ -1,6 +1,7 @@
 #include <ElectroSim/Proton.hpp>
 
-Proton::Proton(int x, int y, int radius) : Particle(x, y, radius) {
+Proton::Proton(int x, int y, float radius) : Particle(x, y, radius) {
+    // Protons have a positive charge
     mCharge = CHARGE;
 }
